Moved LoginForm into Formik/Forms directory.
