+++ src/components/Formik/Input/input.scss
@@ -1,7 +1,5 @@
 @import '../../../utils/scss/constants';
 
-
-
 .error{
   color: $color-redish-five;
   display: flex;
