+++ src/pages/LoginPage/LoginPage.tsx
@@ -1,9 +1,9 @@
 import './_login_page.scss';
 import {Logo} from '../../components/Logo/Logo';
-import {LoginForm} from "../../components/LoginForm/LoginForm";
+import {LoginForm} from "../../components/Formik/Forms/LoginForm";
 
-type Props = {};
-export const LoginPage = (props: Props) => {
+
+export const LoginPage = () => {
   return (
     <div className='login-page'>
       <div className='login-page__container'>
